+++ src/unicode_tests.zig
@@ -185,13 +185,13 @@ test "Segmentation GraphemeIterator" {
         defer all_bytes.deinit();
 
         var graphemes = std.mem.splitSequence(u8, line, " ÷ ");
-        var bytes_index: u32 = 0;
+        var bytes_index: usize = 0;
 
         while (graphemes.next()) |field| {
             var code_points = std.mem.splitScalar(u8, field, ' ');
             var cp_buf: [4]u8 = undefined;
-            var cp_index: u32 = 0;
-            var gc_len: u8 = 0;
+            var cp_index: usize = 0;
+            var gc_len: usize = 0;
 
             while (code_points.next()) |code_point| {
                 if (std.mem.eql(u8, code_point, "×")) continue;
